Support 'extendedKeyUsage' from 286 for prod vs dev code signing.
We do so by adding a new 'purpose' field to the KeySpec. This field
corresponds to a v3 extension section in the openssl.cnf. When
generating certs for KeySpecs (always the root CA keys) we use the
purpose field to determine the appropriate openssl.cnf config section to
set v3 extensions.