pass MODULE_PATH as a parameter & env as PKCS11_PATH
Nixos puts stuff in a special nixos location and we can do stuff to bind
mount things into the places we expect them. Still it's better for this
tool to be flexible. This commit adds a new option to the `ca
initialize` command that allows the caller to pass in the location of
the YubiHSM PKCS#11 module. It can also be provided through the
environment via OKS_PKCS11_PATH.