Increase timeout from 2 minutes to 5.
I just had the HSM timeout while generating an RSA4k key even with a 4
minute timeout. I'm tempted to make this timemout infinite but we need
some way to know when we need intervention in order for the ceremony to
proceed. 5 minutes seems conservative enough while still allowing for a
meaningful timeout.