Add cli tool to create a dcsr spec from the lpc55-sign DCSR struct.
Like a CSR, the DCSR produced by lpc55-sign is missing information that
we need in order to process it. For CSRs our solution was to create a
JSON doc that is used to associate the CSR with the label for the key we
want to certify. We do something similar for DCSRs but we require
additonal data.

Like the CSR spec, we include the label of the key used to sign the
debug credential. Creating a debug credentials also require access to
the trust anchors that are used to create the RKTH entries and so
include labels for these keys as well. Finally the DCSR produced by
`lpc55-sign` is itself serializable to JSON so unlike the CSR that we
include as a PEM string, the DCSR JSON is included in the larger DCSR
spec file.