Move code to dump secrets to output media into a module.
This refactor gets the code responsible for dumping secrets to the
printer out of the HSM module. Additionally this moves all code driving
user interaction out of the HSM module.

This commit also defines a new type for the Feldman Verifier. This is
useful for passing things around w/o having to specify a pile of
generics. This uncovered a bug on the `verify_shares` test case that
required we regenerate some test data.